refactor(repositories): correct return types and add outer-join test

+++ backend/repositories/subscription.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from decimal import Decimal
 
 from sqlalchemy import func, or_, select
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -47,7 +46,7 @@ class SubscriptionRepository(BaseRepository[Subscription]):
         )
         return result.scalar() or 0
 
-    async def sum_saved_for_user(self, user_id: int) -> Decimal:
+    async def sum_saved_for_user(self, user_id: int) -> int:
         result = await self._db.execute(
             select(func.coalesce(func.sum(Appointment.originalPrice), 0)).where(
                 Appointment.userId == user_id,
